Put migrations in separate file
